+++ app/src/main/java/pl/krystiankaniowski/weatherapp/data/openweathermap/model/Snow.kt
@@ -8,6 +8,6 @@ class Snow {
     /** Snow volume for the last 3 hours */
     @SerializedName("3h")
     @Expose
-    val _3h:Int? = null
+    val _3h:Float? = null
 
 }
